+++ ksafe/src/iosTest/kotlin/eu/anifantakis/lib/ksafe/IosKSafeTest.kt
@@ -10,7 +10,7 @@ import kotlin.uuid.Uuid
 class IosKSafeTest : KSafeTest() {
     
     @OptIn(ExperimentalUuidApi::class)
-    override fun createKSafe(fileName: String?): KSafe {
+    override fun newKSafe(fileName: String?): KSafe {
         // If no filename is provided, generate a unique one for each test
         // to avoid DataStore conflicts on iOS
         val actualFileName = fileName ?: Uuid.random().toString()